[MIG] sale_margin_sync: Migration to 19.0
Port sale_margin_sync from 18.0 to 19.0.

- Adapt sale_margin_sync to Odoo 19.0's removal of stock.valuation.layer:
  inherit stock.move and sync the sale line purchase_price from move.value.
- Adapt sale.order.line test data to the product_uom -> product_uom_id rename.
- Adjust tests to set move.value as total cost (unit cost * valued quantity)
  now that stock.valuation.layer.unit_cost no longer exists.
