+++ sale_margin_sync/__manifest__.py
@@ -3,7 +3,7 @@
 {
     "name": "Sale margin sync",
     "summary": "Recompute sale margin when stock move cost price is changed",
-    "version": "18.0.1.0.0",
+    "version": "19.0.1.0.0",
     "category": "Sales",
     "website": "https://github.com/OCA/margin-analysis",
     "author": "Tecnativa, Odoo Community Association (OCA)",
