+++ sale_margin_sync/i18n/es.po
@@ -4,7 +4,7 @@
 #
 msgid ""
 msgstr ""
-"Project-Id-Version: Odoo Server 16.0\n"
+"Project-Id-Version: Odoo Server 19.0\n"
 "Report-Msgid-Bugs-To: \n"
 "PO-Revision-Date: 2023-09-02 18:17+0000\n"
 "Last-Translator: Ivorra78 <informatica@totmaterial.es>\n"
@@ -17,6 +17,6 @@ msgstr ""
 "X-Generator: Weblate 4.17\n"
 
 #. module: sale_margin_sync
-#: model:ir.model,name:sale_margin_sync.model_stock_valuation_layer
-msgid "Stock Valuation Layer"
-msgstr "Nivel de valoración de existencias"
+#: model:ir.model,name:sale_margin_sync.model_stock_move
+msgid "Stock Move"
+msgstr "Movimiento de stock"
