+++ sale_margin_sync/i18n/it.po
@@ -4,7 +4,7 @@
 #
 msgid ""
 msgstr ""
-"Project-Id-Version: Odoo Server 16.0\n"
+"Project-Id-Version: Odoo Server 19.0\n"
 "Report-Msgid-Bugs-To: \n"
 "PO-Revision-Date: 2024-01-29 17:38+0000\n"
 "Last-Translator: mymage <stefano.consolaro@mymage.it>\n"
@@ -17,6 +17,6 @@ msgstr ""
 "X-Generator: Weblate 4.17\n"
 
 #. module: sale_margin_sync
-#: model:ir.model,name:sale_margin_sync.model_stock_valuation_layer
-msgid "Stock Valuation Layer"
-msgstr "Livello valutazione magazzino"
+#: model:ir.model,name:sale_margin_sync.model_stock_move
+msgid "Stock Move"
+msgstr "Movimento di magazzino"
