+++ sale_margin_sync/i18n/sale_margin_sync.pot
@@ -4,7 +4,7 @@
 #
 msgid ""
 msgstr ""
-"Project-Id-Version: Odoo Server 18.0\n"
+"Project-Id-Version: Odoo Server 19.0\n"
 "Report-Msgid-Bugs-To: \n"
 "Last-Translator: \n"
 "Language-Team: \n"
@@ -14,6 +14,6 @@ msgstr ""
 "Plural-Forms: \n"
 
 #. module: sale_margin_sync
-#: model:ir.model,name:sale_margin_sync.model_stock_valuation_layer
-msgid "Stock Valuation Layer"
+#: model:ir.model,name:sale_margin_sync.model_stock_move
+msgid "Stock Move"
 msgstr ""
