+++ sale_margin_sync/i18n/tr.po
@@ -4,7 +4,7 @@
 #
 msgid ""
 msgstr ""
-"Project-Id-Version: Odoo Server 18.0\n"
+"Project-Id-Version: Odoo Server 19.0\n"
 "Report-Msgid-Bugs-To: \n"
 "Last-Translator: Automatically generated\n"
 "Language-Team: none\n"
@@ -15,6 +15,6 @@ msgstr ""
 "Plural-Forms: nplurals=2; plural=n != 1;\n"
 
 #. module: sale_margin_sync
-#: model:ir.model,name:sale_margin_sync.model_stock_valuation_layer
-msgid "Stock Valuation Layer"
+#: model:ir.model,name:sale_margin_sync.model_stock_move
+msgid "Stock Move"
 msgstr ""
