+++ sale_margin_sync/models/__init__.py
@@ -1,3 +1,3 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 
-from . import stock_valuation_layer
+from . import stock_move
