+++ setup/_metapackage/pyproject.toml
@@ -1,10 +1,11 @@
 [project]
 name = "odoo-addons-oca-margin-analysis"
-version = "19.0.20260802.0"
+version = "19.0.20260818.0"
 dependencies = [
     "odoo-addon-account_invoice_margin==19.0.*",
     "odoo-addon-account_invoice_margin_sale==19.0.*",
     "odoo-addon-product_standard_margin==19.0.*",
+    "odoo-addon-sale_margin_sync==19.0.*",
 ]
 classifiers=[
     "Programming Language :: Python",
